refactor: clean authentication/otp
