+++ src/apps/backend/modules/authentication/errors.py
@@ -1,6 +1,11 @@
 from modules.authentication.types import AccessTokenErrorCode
 from modules.error.custom_errors import AppError
 from modules.authentication.types import PasswordResetTokenErrorCode
+from modules.error.custom_errors import AppError
+from modules.authentication.types import PasswordResetTokenErrorCode
+
+from modules.error.custom_errors import AppError
+from modules.authentication.types import OTPErrorCode
 
 
 class AccessTokenInvalidError(AppError):
@@ -29,10 +34,6 @@ class InvalidAuthorizationHeaderError(AppError):
     def __init__(self, message: str) -> None:
         super().__init__(code=AccessTokenErrorCode.INVALID_AUTHORIZATION_HEADER, http_status_code=401, message=message)
 
-from modules.error.custom_errors import AppError
-from modules.authentication.types import PasswordResetTokenErrorCode
-
-
 class PasswordResetTokenNotFoundError(AppError):
 
     def __init__(self) -> None:
@@ -40,4 +41,26 @@ class PasswordResetTokenNotFoundError(AppError):
             code=PasswordResetTokenErrorCode.PASSWORD_RESET_TOKEN_NOT_FOUND,
             http_status_code=404,
             message=f"System is unable to find a token with this account",
+        )
+
+class OTPIncorrectError(AppError):
+    def __init__(self) -> None:
+        super().__init__(
+            code=OTPErrorCode.INCORRECT_OTP, http_status_code=400, message="Please provide the correct OTP to login."
+        )
+
+
+class OTPExpiredError(AppError):
+    def __init__(self) -> None:
+        super().__init__(
+            code=OTPErrorCode.OTP_EXPIRED,
+            http_status_code=400,
+            message="The OTP has expired. Please request a new OTP.",
+        )
+
+
+class OTPRequestFailedError(AppError):
+    def __init__(self) -> None:
+        super().__init__(
+            code=OTPErrorCode.REQUEST_FAILED, http_status_code=400, message="Please provide a valid phone number."
         )
