+++ src/apps/backend/modules/authentication/types.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 from typing import Union
-
+from enum import StrEnum
 from modules.account.types import PhoneNumber
 
 
@@ -60,3 +60,35 @@ class CreatePasswordResetTokenParams:
 @dataclass(frozen=True)
 class PasswordResetTokenErrorCode:
     PASSWORD_RESET_TOKEN_NOT_FOUND: str = "PASSWORD_RESET_TOKEN_ERR_01"
+
+@dataclass(frozen=True)
+class OTPStatus(StrEnum):
+    EXPIRED: str = "EXPIRED"
+    PENDING: str = "PENDING"
+    SUCCESS: str = "SUCCESS"
+
+
+@dataclass(frozen=True)
+class OTP:
+    id: str
+    otp_code: str
+    phone_number: PhoneNumber
+    status: str
+
+
+@dataclass(frozen=True)
+class OTPErrorCode:
+    INCORRECT_OTP: str = "OTP_ERR_01"
+    OTP_EXPIRED: str = "OTP_ERR_02"
+    REQUEST_FAILED: str = "OTP_ERR_03"
+
+
+@dataclass(frozen=True)
+class CreateOTPParams:
+    phone_number: PhoneNumber
+
+
+@dataclass(frozen=True)
+class VerifyOTPParams:
+    otp_code: str
+    phone_number: PhoneNumber
